Completion of tech test
As a client I want to be able to view information about all the fund managers on one page so that users are able to view this information in a simple and responsive design.

For this project I used JSON functions in JQuery to get all the data from the api, I was able to fix my previous issue that did not load all the managers in alphabetical order by using a sort function once every new manager was collected.